refactor: use pure extension lifecycle hooks instead of manual registrations
