+++ features/forms/ui/editor/form-editor.tsx
@@ -32,7 +32,6 @@ import {
   useDataLists,
   useDataListsLoader,
 } from "@/lib/survey-features/data-lists";
-import { useBlindSearchTagbox } from "@/lib/survey-features/blind-search-tagbox";
 import { CreateCustomQuestionRequest } from "@/services/api";
 import "ace-builds/src-noconflict/ace";
 import "ace-builds/src-noconflict/ext-searchbox";
@@ -215,7 +214,6 @@ function FormEditor({
   } = useQuestionLoops();
   const { initGlobals: initDataListsGlobals, setAvailableDataLists } =
     useDataLists();
-  const { initGlobals: initBlindSearchTagboxGlobals } = useBlindSearchTagbox();
   const {
     dataLists,
     isLoading: isDataListsLoading,
@@ -532,7 +530,6 @@ function FormEditor({
         initQuestionLoopsGlobals();
         initFormDiagnosticsGlobals();
         initDataListsGlobals();
-        initBlindSearchTagboxGlobals();
         const newCreator = new SurveyCreator(creatorOptions);
         const resolvedTheme = resolveCreatorThemeCssVariables(
           creatorThemeRef.current,
@@ -608,7 +605,6 @@ function FormEditor({
     bindFormDiagnostics,
     initAnyAnsweredGlobals,
     initDataListsGlobals,
-    initBlindSearchTagboxGlobals,
     initFormDiagnosticsGlobals,
     initQuestionLoopsGlobals,
   ]);
