+++ lib/survey-features/blind-search-tagbox/infrastructure/blind-search-tagbox.extension.ts
@@ -3,8 +3,7 @@ import {
   registerChoicesLazyLoadCompletedHandler,
   registerChoicesLazyLoadGuard,
 } from "@/lib/survey-features/infrastructure/choices-lazy-load-guards";
-import type { SurveyCreatorModel } from "survey-creator-core";
-import type { SurveyInstanceCreatedEvent } from "survey-creator-core";
+import { bindBlindSearchToCreator } from "./creator-bindings";
 import { registerBlindSearchTagboxGlobals } from "./registry";
 import {
   applyBlindSearchOtherFallbackAfterLazyLoad,
@@ -12,55 +11,17 @@ import {
 } from "./survey-bindings";
 import { shouldSuppressChoices } from "../use-cases/blind-search-state";
 
-const BLIND_SEARCH_CREATOR_BOUND_KEY = "__endatixBlindSearchCreatorBound";
 const BLIND_SEARCH_LAZY_LOAD_GUARD_ID = "blind-search-tagbox";
-const DESIGNER_TAB_SURVEY_AREA = "designer-tab";
 
-function bindBlindSearchToCreator(creator: SurveyCreatorModel): () => void {
-  const creatorWithFlags = creator as SurveyCreatorModel & Record<string, unknown>;
-  if (creatorWithFlags[BLIND_SEARCH_CREATOR_BOUND_KEY]) {
-    return () => {};
-  }
-  creatorWithFlags[BLIND_SEARCH_CREATOR_BOUND_KEY] = true;
-
-  const creatorSurveyDisposers = new Map<string, () => void>();
-
-  const bindSurveyForCreatorArea = (
-    area: string,
-    survey: SurveyInstanceCreatedEvent["survey"],
-  ) => {
-    const previousDispose = creatorSurveyDisposers.get(area);
-    previousDispose?.();
-
-    const dispose = bindBlindSearchToSurvey(survey);
-    creatorSurveyDisposers.set(area, dispose);
-  };
-
-  const handleSurveyInstanceCreated = (
-    _: unknown,
-    options: SurveyInstanceCreatedEvent,
-  ) => {
-    if (options.area === "property-grid") {
-      return;
-    }
-
-    bindSurveyForCreatorArea(options.area, options.survey);
-  };
-
-  creator.onSurveyInstanceCreated.add(handleSurveyInstanceCreated);
-
-  if (creator.survey) {
-    bindSurveyForCreatorArea(DESIGNER_TAB_SURVEY_AREA, creator.survey);
-  }
-
-  return () => {
-    creator.onSurveyInstanceCreated.remove(handleSurveyInstanceCreated);
-    creatorSurveyDisposers.forEach((dispose) => dispose?.());
-    creatorSurveyDisposers.clear();
-    creatorWithFlags[BLIND_SEARCH_CREATOR_BOUND_KEY] = false;
-  };
-}
-
+/**
+ * Survey extension entry point. All install logic lives here — do not wire this
+ * feature from form-editor via initGlobals / bindToCreator hooks.
+ *
+ * Lifecycle:
+ * - onInit: Serializer metadata + cross-feature guard registries (runs when extension is selected)
+ * - onCreatorReady: designer preview and Creator survey instances
+ * - onModelReady: respondent / standalone Model surfaces
+ */
 const blindSearchTagboxExtension: ExtensionModule = {
   onInit: () => {
     registerBlindSearchTagboxGlobals();
@@ -81,4 +42,4 @@ const blindSearchTagboxExtension: ExtensionModule = {
   },
 };
 
-export { blindSearchTagboxExtension, bindBlindSearchToCreator, bindBlindSearchToSurvey };
+export { blindSearchTagboxExtension };
